improve log
